requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.2
pandas==2.2.1
scikit-learn==1.4.1.post1
